# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Software Department Portfolio Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sprints:
    get:
      operationId: listSprints
      tags: [sprints]
      summary: List sprints, optionally filtered by cycle
      parameters:
        - name: cycleId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of sprints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sprint"
    post:
      operationId: createSprint
      tags: [sprints]
      summary: Create a sprint (editor only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSprintBody"
      responses:
        "201":
          description: Created sprint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sprint"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sprints/{id}:
    patch:
      operationId: updateSprint
      tags: [sprints]
      summary: Update a sprint (editor only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSprintBody"
      responses:
        "200":
          description: Updated sprint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sprint"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSprint
      tags: [sprints]
      summary: Delete a sprint (editor only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users:
    get:
      operationId: listUsers
      tags: [auth]
      summary: List all managed users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createUser
      tags: [auth]
      summary: Create a managed user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [auth]
      summary: Update a managed user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteUser
      tags: [auth]
      summary: Delete a managed user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User deleted
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    AppUser:
      type: object
      properties:
        id:
          type: integer
        clerkUserId:
          type: ["string", "null"]
        email:
          type: string
        role:
          type: string
          enum: [admin, guest]
        team:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkUserId
        - email
        - role
        - team
        - createdAt
    CreateUserBody:
      type: object
      properties:
        email:
          type: string
        clerkUserId:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, guest]
        team:
          type: ["string", "null"]
      required:
        - email
        - role
    UpdateUserBody:
      type: object
      properties:
        role:
          type: string
          enum: [admin, guest]
        team:
          type: ["string", "null"]
...
    Sprint:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        cycleId:
          type: integer
        sprintNumber:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - cycleId
        - sprintNumber
        - startDate
        - endDate
        - createdAt
    CreateSprintBody:
      type: object
      properties:
        name:
          type: string
        cycleId:
          type: integer
        sprintNumber:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
      required:
        - name
        - cycleId
        - sprintNumber
        - startDate
        - endDate
    UpdateSprintBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        cycleId:
          type: ["integer", "null"]
        sprintNumber:
          type: ["integer", "null"]
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date
...